% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourscheme.R
\name{add_other}
\alias{add_other}
\title{Add an "Other" observation to a relative abundance table.}
\usage{
add_other(rel_abund_tb)
}
\arguments{
\item{rel_abund_tb}{A relative abundance table in tibble format.}
}
\value{
A tibble.
}
\description{
Add an "Other" observation to a relative abundance table.
}
\examples{
rel_abund_phy(physeq1) \%>\% add_other()
}
