% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourscheme.R
\name{all_taxa}
\alias{all_taxa}
\title{Return a character vector of unique taxa from a relative abundance table.}
\usage{
all_taxa(rel_abund_tb)
}
\arguments{
\item{rel_abund_tb}{A relative abundance table in tibble format.}
}
\value{
A vector arranged by abundance.
}
\description{
Return a character vector of unique taxa from a relative abundance table.
}
\examples{
rel_abund_phy(phy = physeq1) \%>\%
    all_taxa()
}
