% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_sample_by_taxa}
\alias{arrange_sample_by_taxa}
\title{Arrange samples by the most abundant taxa.}
\usage{
arrange_sample_by_taxa(rel_abund_tb)
}
\arguments{
\item{rel_abund_tb}{A relative abundance table in tibble form.}
}
\value{
A tibble with ordered samples.
}
\description{
Arrange samples by the most abundant taxa.
}
\examples{
rel_abund_phy(phy = physeq1) \%>\%
     arrange_sample_by_taxa()
}
