% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_taxa}
\alias{arrange_taxa}
\title{Arrange taxa by abundance.}
\usage{
arrange_taxa(rel_abund_tb, pooled = "top", order = "bottom")
}
\arguments{
\item{rel_abund_tb}{A relative abundance table in tibble form.}

\item{pooled}{The order of pooled threshold, either "top" or "bottom".}

\item{order}{The order of taxa, either "top" or "bottom".}
}
\value{
A tibble with ordered taxon.
}
\description{
Arrange taxa by abundance.
}
\examples{
rel_abund_phy(phy = physeq) \%>\%
     arrange_taxa(pooled = "bottom")
}
