% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_var_abund}
\alias{arrange_var_abund}
\title{Arrange a variable by its grouped relative abundance}
\usage{
arrange_var_abund(rel_abund_tb, var = "sample_id", flip = FALSE)
}
\arguments{
\item{rel_abund_tb}{A relative abundance table in tibble form.}

\item{var}{The variable to group by. Default is "sample_id".}

\item{flip}{Logical, whether to flip the variable.}
}
\value{
A tibble with sorted variable.
}
\description{
Arrange a variable by its grouped relative abundance
}
\examples{
rel_abund_phy(physeq) \%>\%
    arrange_var_abund()
}
