% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{asv_data_qiime}
\alias{asv_data_qiime}
\title{Grab asv table from a .qza file and format to tibble.}
\usage{
asv_data_qiime(qza)
}
\arguments{
\item{qza}{Path to .qza file containing otu counts.}
}
\value{
A tibble.
}
\description{
Grab asv table from a .qza file and format to tibble.
}
\examples{
fpath <- system.file("extdata/qiime", "table-dada2.qza", package = "bubbler")
asv_data_qiime(fpath)
}
