% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tsv.R
\name{asv_data_tsv}
\alias{asv_data_tsv}
\title{Grab asv table from a tsv file and format to tibble}
\usage{
asv_data_tsv(tsv)
}
\arguments{
\item{tsv}{A character vector specifying the path to seqtab.tsv}
}
\value{
A tibble.
}
\description{
Grab asv table from a tsv file and format to tibble
}
\examples{
fpath <- system.file("extdata", "seqtab.tsv", package = "bubbler")
asv_data_tsv(fpath)
}
