% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{asv_tree}
\alias{asv_tree}
\title{Generate a ggtree object based on the distance metric, from asv data}
\usage{
asv_tree(asv_data, method = "bray")
}
\arguments{
\item{asv_data}{An asv_data_<qiime, tsv, phy, bracken> tibble.}

\item{method}{Distance metric to be used. Default is "bray".}
}
\value{
A ggtree object.
}
\description{
Generate a ggtree object based on the distance metric, from asv data
}
\examples{
counts_q <- system.file("extdata", "qiime", "table-dada2.qza", package = "bubbler")
asv_data_qiime(counts_q) \%>\%
    asv_tree()
}
