% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bubble_plot}
\alias{bubble_plot}
\title{Generate a ggplot2 bubble plot}
\usage{
bubble_plot(rel_abund_tb, x_var = "sample_id", color = NULL, italics = FALSE)
}
\arguments{
\item{rel_abund_tb}{A relative abundance table in tibble format.}

\item{x_var}{The variable to used on the x-axis.}

\item{color}{The color.}

\item{italics}{Logical. Whether to use italics or not.}
}
\value{
A ggplot
}
\description{
Generate a ggplot2 bubble plot
}
\examples{
rel_abund_phy(physeq) \%>\% bubble_plot()
}
