% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{choose_n_taxa}
\alias{choose_n_taxa}
\title{Choose how many taxa to display when plotting}
\usage{
choose_n_taxa(rel_abund_tab, n_taxa = 8)
}
\arguments{
\item{rel_abund_tab}{A rel_abund table in tibble format.}

\item{n_taxa}{An integer vector for the number of taxa to display.}
}
\value{
A numeric vector.
}
\description{
Generates a numeric vector as the threshold to display n taxa.
}
\examples{
choose_n_taxa(rel_abund_phy(physeq1), n_taxa = 8)
}
