% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{choose_samples}
\alias{choose_samples}
\title{Filter a rel_abund table for a selection of samples}
\usage{
choose_samples(rel_abund, smp_selection)
}
\arguments{
\item{rel_abund}{A rel_abund table in tibble format.}

\item{smp_selection}{A character vector of the selected samples.}
}
\value{
A tibble.
}
\description{
Filter a rel_abund table for a selection of samples
}
\examples{
choose_samples(rel_abund, c("Smp1", "Smp2", "Smp3"))
}
