% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{choose_samples_asv_phy}
\alias{choose_samples_asv_phy}
\title{Select samples prior to calculating rel_abund}
\usage{
choose_samples_asv_phy(phy, smp_selection)
}
\arguments{
\item{phy}{A phyloseq object .}

\item{smp_selection}{A character vector of sample selections.}
}
\value{
A tibble.
}
\description{
Select samples prior to calculating rel_abund
}
