% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose.R
\name{choose_taxa_level}
\alias{choose_taxa_level}
\title{Filter a rel_abund table at a specific taxonomic level}
\usage{
choose_taxa_level(rel_abund, taxon_level = "Phylum")
}
\arguments{
\item{rel_abund}{A rel_abund table in tibble format.}

\item{taxon_level}{A character vector of the taxonomic level.}
}
\value{
A tibble.
}
\description{
Filter a rel_abund table at a specific taxonomic level
}
\examples{
choose_taxa_level(rel_abund, "Phylum")
}
