% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourscheme.R
\name{combine_dataframes}
\alias{combine_dataframes}
\title{Combine data.frames that are in a list.}
\usage{
combine_dataframes(dataframes, by_var)
}
\arguments{
\item{dataframes}{List of dataframes.}

\item{by_var}{Variable to join by.}
}
\value{
A joined dataframe.
}
\description{
Combine data.frames that are in a list.
}
\examples{
df1 <- data.frame(id = 1:3, value = c("A", "B", "C"))
df2 <- data.frame(id = 2:4, value = c("D", "E", "F"))
df3 <- data.frame(id = 1:4, value = c("G", "H", "I", "J"))
dataframes <- list(df1, df2, df3)
combine_dataframes(dataframes, by_var = "value")
}
