% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{detect_threshold_vec}
\alias{detect_threshold_vec}
\title{Detect the label used in a character vector.}
\usage{
detect_threshold_vec(vector)
}
\arguments{
\item{vector}{The character vector of taxon.}
}
\value{
A character vector of length one.
}
\description{
Detect the label used in a character vector.
}
\examples{
rel_abund_phy(physeq) \%>\% all_taxa() \%>\% dplyr::pull(taxon) \%>\% detect_threshold_vec()
}
