% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{.pool_taxon_thresh}
\alias{.pool_taxon_thresh}
\title{Helper function 1 to pool taxa below threshold}
\usage{
.pool_taxon_thresh(rel_abund, threshold = 0.2)
}
\arguments{
\item{rel_abund}{A rel_abund table in tibble format.}

\item{threshold}{A numeric vector for the threshold.}
}
\value{
A tibble.
}
\description{
Summarise the pooled group as a logical vector, also summarise
the mean rel_abund values to be used for ordering.
}
\examples{
.pool_taxon_tresh(rel_abund, threshold)
}
