% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{.set_taxon_threshold}
\alias{.set_taxon_threshold}
\title{Helper function 2 to pool taxa below threshold}
\usage{
.set_taxon_threshold(rel_abund_pool, taxon, threshold = 0.2)
}
\arguments{
\item{rel_abund_pool}{A pooled rel_abund table.}

\item{taxon}{The taxon column.}
}
\value{
A tibble.
}
\description{
Pool the taxon that are below the threshold
}
\examples{
.set_taxon_threshold(rel_abund_pool, taxon)
}
