% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourscheme.R
\name{extract_unique_taxa}
\alias{extract_unique_taxa}
\title{Get the unique taxa from the combined data.frame.}
\usage{
extract_unique_taxa(...)
}
\arguments{
\item{...}{The dataframes to join.}
}
\value{
A character vector of the unique taxon from the combined data.frames.
}
\description{
Get the unique taxa from the combined data.frame.
}
\examples{
asv <- system.file("extdata/tsv", "seqtab.tsv", package = "bubbler")
taxa <- system.file("extdata/tsv", "taxa.tsv", package = "bubbler")
rel_abund <- rel_abund_tsv(asv, taxa)
rel_abund_l <- rel_abund \%>\%
    subset_high_low(n = 10)
rel_abund_h <- rel_abund \%>\%
    subset_high_low(n = 50, flip = TRUE)
extract_unique_taxa(rel_abund_l, rel_abund_h)
}
