% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colourscheme.R
\name{global_colour_scheme}
\alias{global_colour_scheme}
\title{Set a global colour scheme for a series of plots.}
\usage{
global_colour_scheme(all_taxa, unique_taxa)
}
\arguments{
\item{all_taxa}{All taxa from the initial relative abundance table, as a character vector.}

\item{unique_taxa}{The unique taxa among your subsetted/pooled relative abundance table.}
}
\value{
A named vector.
}
\description{
Set a global colour scheme for a series of plots.
}
\examples{
asv <- system.file("extdata/tsv", "seqtab.tsv", package = "bubbler")
taxa <- system.file("extdata/tsv", "taxa.tsv", package = "bubbler")
rel_abund <- rel_abund_tsv(asv, taxa) \%>\% add_other()
rel_abund_l <- rel_abund \%>\%
    subset_high_low(n = 10)
rel_abund_h <- rel_abund \%>\%
    subset_high_low(n = 50, flip = TRUE)
unique_taxa <- extract_unique_taxa(rel_abund_l, rel_abund_h)
all_taxa <- rel_abund \%>\% all_taxa()
global_colour_scheme(all_taxa, unique_taxa)
}
