% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{meta_data_phy}
\alias{meta_data_phy}
\title{Grab metadata from a physeq object and format to tibble}
\usage{
meta_data_phy(phy)
}
\arguments{
\item{phy}{A phyloseq object containig sample_data}
}
\value{
A tibble.
}
\description{
Grab metadata from a physeq object and format to tibble
}
\examples{
meta_data_phy(physeq1)
}
