% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{meta_data_qiime}
\alias{meta_data_qiime}
\title{Grab metadata table from a .tsv QIIME2-formatted file and format to tibble.}
\usage{
meta_data_qiime(tsv)
}
\arguments{
\item{tsv}{Path to .tsv file in QIIME2 format.}
}
\value{
A tibble.
}
\description{
Grab metadata table from a .tsv QIIME2-formatted file and format to tibble.
}
\examples{
fpath <- system.file("extdata/qiime", "sample-metadata.tsv", package = "bubbler")
meta_data_qiime(fpath)
}
