% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_tsv.R
\name{meta_data_tsv}
\alias{meta_data_tsv}
\title{Grab metadata table from a tsv file and format to tibble}
\usage{
meta_data_tsv(tsv)
}
\arguments{
\item{tsv}{A character vector specifying the path to metadata.tsv}
}
\value{
A tibble.
}
\description{
Grab metadata table from a tsv file and format to tibble
}
\examples{
fpath <- system.file("extdata", "metadata.tsv", package = "bubbler")
meta_data_tsv(fpath)
}
