% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{physeq}
\alias{physeq}
\title{Phyloseq object for bubbler}
\description{
A 100 sample by 20 taxa dataset for experimentation. Counts were generated from a log-normal
distribution and taxonomic information was randomly sampled from the RDP database.
}
\author{
Zjardyn Liera-Hood \email{zlieraho@uwaterloo.ca}
}
\keyword{data}
