% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_rel_abund.R
\docType{data}
\name{physeq1}
\alias{physeq1}
\title{This is data to be included in my package}
\description{
This is data to be included in my package
}
\author{
Zjardyn Liera-Hood \email{zlieraho@uwaterloo.ca}
}
\keyword{data}
