% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{pool_taxa}
\alias{pool_taxa}
\title{Pool taxa according to threshold}
\usage{
pool_taxa(
  rel_abund_tab,
  threshold,
  n_taxa = 12,
  keep_metadata = FALSE,
  label = TRUE
)
}
\arguments{
\item{rel_abund_tab}{A rel_abund table in tibble format.}

\item{threshold}{A numeric vector for the threshold.}

\item{n_taxa}{The number of taxa to display.}

\item{keep_metadata}{Logical. Whether to keep metadata or not.}

\item{label}{Logical. Whether to show threshold in label.}
}
\value{
A tibble.
}
\description{
Applies a threshold and pools any taxa below this threshold, across samples
and optionally, across a variable.
}
\examples{
rel_abund_phy(physeq) \%>\% pool_taxa()
}
