% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{read_bracken_file}
\alias{read_bracken_file}
\title{Grab bracken data and format to tibble.}
\usage{
read_bracken_file(filepath)
}
\arguments{
\item{filepath}{Path to bracken .txt files.}
}
\value{
A tibble.
}
\description{
Grab bracken data and format to tibble.
}
\examples{
fpath <- system.file("extdata/bracken",
"7_S79_kraken2_report_bracken_output.txt", package = "bubbler")
read_bracken_file(fpath)
}
