% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qiime2R.R
\name{read_qza}
\alias{read_qza}
\title{read qiime2 artifacts (.qza)}
\usage{
read_qza(file, tmp, rm)
}
\arguments{
\item{file}{path to the input file, ex: file="~/data/moving_pictures/table.qza"}

\item{tmp}{a temporary directory that the object will be decompressed to (default="tempdir()")}

\item{rm}{should the decompressed object be removed at completion of function (T/F default=TRUE)}
}
\value{
a named list.
}
\description{
extracts embedded data and object metadata into an R session
}
\examples{
fpath <- system.file("extdata/qiime", "table-dada2.qza", package = "bubbler")
read_qza(fpath)
}
