% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_rel_abund.R
\name{rel_abund}
\alias{rel_abund}
\title{Generate a relative abundance table from a phyloseq object, summed across samples}
\usage{
rel_abund(phy)
}
\arguments{
\item{phy}{A phyloseq object containing an otu_table and tax_table.}
}
\value{
A tibble.
}
\description{
Generate a relative abundance table from a phyloseq object, summed across samples
}
\examples{
rel_abund(phy = physeq1)
}
