% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_rel_abund.R
\name{rel_abund_bracken}
\alias{rel_abund_bracken}
\title{Generate a relative abundance table using kracken2/bracken data.}
\usage{
rel_abund_bracken(path, remove_human)
}
\arguments{
\item{path}{The filepath to bracken data in .txt format.}

\item{remove_human}{Logical. Choose to remove human samples.}
}
\value{
A tibble.
}
\description{
Generate a relative abundance table using kracken2/bracken data.
}
\examples{
bracken <- system.file("extdata/bracken", package = "bubbler")
rel_abund_bracken(bracken)
}
