% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_rel_abund.R
\name{rel_abund_raw}
\alias{rel_abund_raw}
\title{Generate a relative abundance table in tibble format from raw tsv files.}
\usage{
rel_abund_raw(asv, taxa, meta_data = NULL, var = NULL, taxa_level = "Phylum")
}
\arguments{
\item{asv}{A tsv file path containing an asv table.}

\item{taxa}{A tsv file path containing a taxa table.}

\item{meta_data}{Either NULL, or, optionally, a tsv file path containing a meta_data table.}

\item{var}{A character value of a variable to sum by.}

\item{taxa_level}{A character value specifying the taxa level from Domain to species.}
}
\value{
A tibble.
}
\description{
Generate a relative abundance table in tibble format from raw tsv files.
}
\examples{
asv <- system.file("extdata", "seqtab.tsv", package = "bubbler")
taxa <- system.file("extdata", "taxa.tsv", package = "bubbler")
meta_data <- system.file("extdata", "metadata.tsv", package = "bubbler")
rel_abund_raw(asv, taxa, meta_data)
}
