% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_rel_abund.R
\name{rel_abund_var}
\alias{rel_abund_var}
\title{Generate a relative abundance table from a phyloseq object, summed across a variable}
\usage{
rel_abund_var(phy, var)
}
\arguments{
\item{phy}{A phyloseq object containing an otu_table, tax_table, and sample_data.}

\item{var}{A character vector containing the variable present in sample_data.}
}
\value{
A tibble.
}
\description{
Generate a relative abundance table from a phyloseq object, summed across a variable
}
\examples{
rel_abund_var(phy = physeq1, var = "Location")
}
