% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{show_top_taxa}
\alias{show_top_taxa}
\title{Show the top taxa of a rel_abund table}
\usage{
show_top_taxa(rel_abund_tab)
}
\arguments{
\item{rel_abund_tab}{A rel_abund table in tibble format.}
}
\value{
A tibble.
}
\description{
Finds the top taxa and arranges in descending order. Uses max() to
find the maximum taxon.
}
\examples{
show_top_taxa(rel_abund_phy(physeq))
}
