% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_rel_abund.R
\name{sum_rel_abund}
\alias{sum_rel_abund}
\title{Sum the relative abundance across variables.}
\usage{
sum_rel_abund(rel_abund_tb, ...)
}
\arguments{
\item{rel_abund_tb}{A relative abundance table in}

\item{...}{Variables as strings.}
}
\value{
Summaries in tibble format.
}
\description{
Sum the relative abundance across variables.
}
\examples{
rel_abund_phy(physeq) \%>\%
sum_rel_abund("sample_id")
}
