% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{taxa_data_phy}
\alias{taxa_data_phy}
\title{Grab taxa table from a physeq object and format to tibble}
\usage{
taxa_data_phy(phy)
}
\arguments{
\item{phy}{A phyloseq object containing  tax_table}
}
\value{
A tibble.
}
\description{
Grab taxa table from a physeq object and format to tibble
}
\examples{
taxa_data_phy(physeq)
}
