% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{taxa_data_qiime}
\alias{taxa_data_qiime}
\title{Grab taxonomy table from a .qza file and format to tibble.}
\usage{
taxa_data_qiime(qza)
}
\arguments{
\item{qza}{Path to .qza file containing a taxonomy table.}
}
\value{
A tibble.
}
\description{
Grab taxonomy table from a .qza file and format to tibble.
}
\examples{
fpath <- system.file("extdata/qiime", "taxonomy.qza", package = "bubbler")
taxa_data_qiime(fpath)
}
