% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_data.R
\name{taxa_data_tsv}
\alias{taxa_data_tsv}
\title{Grab taxa table from a tsv file and format to tibble}
\usage{
taxa_data_tsv(tsv)
}
\arguments{
\item{tsv}{A character vector specifying the path to seqtab.tsv}
}
\value{
A tibble.
}
\description{
Grab taxa table from a tsv file and format to tibble
}
\examples{
fpath <- system.file("extdata", "taxa.tsv", package = "bubbler")
taxa_data_tsv(fpath)
}
