% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{taxon_italics}
\alias{taxon_italics}
\title{Convert taxon to italics, ignoring threshold, from a relative abundance tibble.}
\usage{
taxon_italics(rel_abund_tb)
}
\arguments{
\item{rel_abund_tb}{A relative abundance table formatted as a tibble.}
}
\value{
A tibble.
}
\description{
Convert taxon to italics, ignoring threshold, from a relative abundance tibble.
}
\examples{
 rel_abund_phy(physeq) \%>\%
    taxon_italics()
}
