% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree.R
\name{tip_order}
\alias{tip_order}
\title{Grab the order of tips (sample_id) to use as levels for plotting.}
\usage{
tip_order(tree)
}
\arguments{
\item{tree}{A ggtree object.}
}
\value{
A vector of level order.
}
\description{
Grab the order of tips (sample_id) to use as levels for plotting.
}
\examples{
counts_q <- system.file("extdata", "qiime", "table-dada2.qza", package = "bubbler")
asv_data_qiime(counts_q) \%>\%
    asv_tree() \%>\%
    tip_order()
}
