% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange.R
\name{arrange_variable}
\alias{arrange_variable}
\title{Arrange a variable by the ordering of another variable}
\usage{
arrange_variable(rel_abund_tb, variable = "sample_id", levels)
}
\arguments{
\item{rel_abund_tb}{A relative abundance table in tibble form.}

\item{variable}{The variable to be sorted.}

\item{levels}{The variable to sort with.}
}
\value{
A tibble with a sorted variable.
}
\description{
Arrange a variable by the ordering of another variable
}
\examples{
rel_abund_phy(phy = physeq, meta_data = TRUE) \%>\%
    arrange_variable(levels = "Location")
}
