% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{bar_plot}
\alias{bar_plot}
\title{Generate a ggplot2 stacked barplot from a relative abundance table.}
\usage{
bar_plot(rel_abund_tb, x_var = "sample_id", position = "stack", width = 1)
}
\arguments{
\item{rel_abund_tb}{A relative abundance table in tibble format.}

\item{x_var}{The x variable for plotting.}

\item{position}{"fill" or "stack", should the plotting area be filled in.}

\item{width}{The width of the bar.}
}
\value{
A ggplot object.
}
\description{
Generate a ggplot2 stacked barplot from a relative abundance table.
}
\examples{
rel_abund_phy(physeq1) \%>\% bar_plot()
}
