% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/threshold.R
\name{pool_taxa}
\alias{pool_taxa}
\title{Pool taxa according to threshold}
\usage{
pool_taxa(rel_abund_tab, threshold)
}
\arguments{
\item{rel_abund_tab}{A rel_abund table in tibble format.}

\item{threshold}{A numeric vector for the threshold.}
}
\value{
A tibble.
}
\description{
Applies a threshold and pools any taxa below this threshold, across samples
and optionally, across a variable.
}
\examples{
\dontrun{pool_taxa(rel_abund_tab, threshold = 0.2, var = "Location")}
}
