% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asv_rel_abund.R
\name{rel_abund}
\alias{rel_abund}
\title{Generate a relative abundance table from a phyloseq object.}
\usage{
rel_abund(phy, taxa_level = "Phylum", meta_data = FALSE)
}
\arguments{
\item{phy}{A phyloseq object containing an otu_table and tax_table.}

\item{taxa_level}{A character value specifying the taxa level from Domain to species.}

\item{meta_data}{A logical value specifying if metadata should be included from the phyloseq object.}
}
\value{
A tibble.
}
\description{
Generate a relative abundance table from a phyloseq object.
}
\examples{
rel_abund(phy = physeq1)
}
